elm_common: webpack.config.js now propagates WEBPACK_* variables
